requirements.txt
pytest==9.0.2
playwright==1.57.0
pytest-playwright==0.7.2
allure-pytest==2.15.3
